requirements.txt
Django==1.4.1
PIL==1.1.7
Pillow==1.7.7
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.13
boto==2.5.2
celery==3.0.9
distribute==0.6.27
dj-database-url==0.2.1
django-celery==3.0.9
django-endless-pagination==1.1
django-simple-captcha==0.3.4
django-storages==1.1.5
gunicorn==0.14.6
indextank==1.0.6
kombu==2.4.6
psycopg2==2.4.5
python-dateutil==1.5
wsgiref==0.1.2
